+++ modules/ext.Verified_Page_History.signMessage/init.js
